--- a/Bencodex.Tests/Types/ListTest.cs
+++ b/Bencodex.Tests/Types/ListTest.cs
@@ -1,14 +1,16 @@
+using System.Collections.Immutable;
 using System.Linq;
 using System.Text;
-using Bencodex.Misc;
 using Bencodex.Types;
 using Xunit;
 using static Bencodex.Misc.ImmutableByteArrayExtensions;
+using static Bencodex.Tests.TestUtils;
 
 namespace Bencodex.Tests.Types
 {
     public class ListTest
     {
+        private static readonly Codec _codec;
         private static List _zero;
         private static List _one;
         private static List _two;
@@ -16,17 +18,42 @@ namespace Bencodex.Tests.Types
 
         static ListTest()
         {
-            _zero = default(List);
-            _one = new List(new IValue[] { default(Null) });
+            _codec = new Codec();
+            _zero = List.Empty;
+            _one = new List(Null.Value);
             _two = new List(new Text[] { "hello", "world" }.Cast<IValue>());
-            _nest = new List(
-                new IValue[]
-                {
-                    default(Null),
-                    _zero,
-                    _one,
-                    _two,
-                }
+            _nest = new List(Null.Value, _zero, _one, _two);
+        }
+
+        [Fact]
+        public void Constructors()
+        {
+            Assert.Equal(
+                _one.Add(_zero).Add(_one).Add(_two),
+                new List(ImmutableArray.Create<IValue>(Null.Value, _zero, _one, _two))
+            );
+
+            Assert.Equal(_zero, new List(Enumerable.Empty<IValue>())
+            );
+            Assert.Equal(
+                new List(ImmutableArray<IValue>.Empty.Add(Null.Value)),
+                new List(Enumerable.Empty<IValue>().Append(Null.Value))
+            );
+            Assert.Equal(
+                new List((Text)"hello", (Text)"world"),
+                new List(new Text[] { "hello", "world" }.Cast<IValue>())
+            );
+            Assert.Equal(
+                _nest,
+                new List(new IValue[] { Null.Value, _zero }.Concat(new IValue[] { _one, _two }))
+            );
+
+            Assert.Equal(_zero, new List());
+            Assert.Equal(_zero.Add(Null.Value), new List(Null.Value));
+            Assert.Equal(_two, new List((Text)"hello", (Text)"world"));
+            Assert.Equal(
+                _one.Add(_zero).Add(_one).Add(_two),
+                new List(Null.Value, _zero, _one, _two)
             );
         }
 
@@ -101,7 +128,7 @@ namespace Bencodex.Tests.Types
 
             // If any element is a list/dict it should be indented
             Assert.Equal("[\n  [],\n]", new List(new IValue[] { _zero }).Inspection);
-            Assert.Equal("[\n  {},\n]", new List(new IValue[] { Dictionary.Empty }).Inspection);
+            Assert.Equal("[\n  {},\n]", new List(Dictionary.Empty).Inspection);
         }
 
         [Fact]
@@ -149,6 +176,33 @@ namespace Bencodex.Tests.Types
             Assert.Equal((Boolean)true, list[3]);
             Assert.Equal(List.Empty, list[4]);
             Assert.Equal(Dictionary.Empty, list[5]);
+        }
+
+        [Fact]
+        public void Encode()
+        {
+            AssertEqual(
+                new byte[] { 0x6c, 0x65 },  // "le"
+                _codec.Encode(_zero)
+            );
+            AssertEqual(
+                new byte[] { 0x6c, 0x65 },  // "le"
+                _codec.Encode(new List())
+            );
+            AssertEqual(
+                new byte[] { 0x6c, 0x65 },  // "le"
+                _codec.Encode(new List(ImmutableList<IValue>.Empty))
+            );
+            AssertEqual(
+                new byte[]
+                {
+                    0x6c, 0x75, 0x35, 0x3a, 0x68, 0x65, 0x6c, 0x6c, 0x6f,
+                    0x75, 0x35, 0x3a, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x65,
+
+                    // "lu5:hellou5:worlde"
+                },
+                _codec.Encode(_two)
+            );
         }
     }
 }