--- a/CHANGES.md
+++ b/CHANGES.md
@@ -7,16 +7,17 @@ Version 0.4.0
 To be released.
 
  -  Added `Bencodex.Types.ValueType` enum type.  [[#50]]
- -  Added `Bencodex.Types.Fingerprint` readonly struct.  [[#50]]
- -  Added `Bencodex.Misc.FingerprintComparer` class.  [[#50]]
+ -  Bencodex values now have their unqiue fingerprints:  [[#50]]
+     -  Added `Bencodex.Types.Fingerprint` readonly struct.
+     -  Added `Bencodex.Misc.FingerprintComparer` class.
+     -  Added `IValue.Fingerprint` property.
+     -  Added `Null.SingletonFingerprint` static readonly field.
+     -  Added `Boolean.TrueFingerprint` static readonly field.
+     -  Added `Boolean.FalseFingerprint` static readonly field.
+     -  Added `List.EmptyFingerprint` static readonly field.
+     -  Added `Dictionary.EmptyFingerprint` static readonly field.
  -  Added `IValue.Type` property.  [[#50]]
- -  Added `IValue.Fingerprint` property.  [[#50]]
  -  Added `IValue.EncodingLength` property.  [[#49], [#50]]
- -  Added `Null.SingletonFingerprint` static readonly field.  [[#50]]
- -  Added `Boolean.TrueFingerprint` static readonly field.  [[#50]]
- -  Added `Boolean.FalseFingerprint` static readonly field.  [[#50]]
- -  Added `List.EmptyFingerprint` static readonly field.  [[#50]]
- -  Added `Dictionary.EmptyFingerprint` static readonly field.  [[#50]]
  -  Added `Bencodex.Misc.ImmutableByteArrayExtensions` static class.  [[#50]]
  -  Replaced `Binary(byte[])` constructor with `Binary(params byte[])`
     constructor.  [[#50]]