--- a/CHANGES.md
+++ b/CHANGES.md
@@ -6,6 +6,7 @@ Version 0.4.0
 
 To be released.
 
+ -  `Bencodex.Types.List` struct became a class.  [[#51]]
  -  Added `Bencodex.Types.ValueType` enum type.  [[#50]]
  -  Bencodex values now have their unqiue fingerprints:  [[#50]]
      -  Added `Bencodex.Types.Fingerprint` readonly struct.