… Bencodex/Types/Dictionary.cs