This patch optimizes `List` and `Dictionary` types by refactoring them.

- Both types are now reference types.
- Every `Dictionary` instance is always sorted as [`ImmutableSortedDictionary<K, V>`][1] is used as its internal data structure.
    - Due to this, I can get rid of its dual internal representations (`Dict` mode & `Pairs` mode) which were quite complicated.

### Benchmarks before this patch

|           Method |             DataFile |      Mean |    Error |    StdDev |
|----------------- |--------------------- |----------:|---------:|----------:|
|  EncodeIntoBytes | /tmp/(...)e.dat [96] | 27.853 ms | 3.870 ms | 0.2121 ms |
| EncodeIntoStream | /tmp/(...)e.dat [96] | 21.257 ms | 4.667 ms | 0.2558 ms |
|  DecodeFromBytes | /tmp/(...)e.dat [96] |  8.122 ms | 8.924 ms | 0.4891 ms |
| DecodeFromStream | /tmp/(...)e.dat [96] |  8.216 ms | 5.635 ms | 0.3089 ms |

|                  Method |         Mean |      Error |    StdDev |
|------------------------ |-------------:|-----------:|----------:|
|    List_AddBinFromEmpty |    44.08 ns |     1.898 ns |  0.104 ns |
|    List_AddTxtFromEmpty |    42.67 ns |     2.461 ns |  0.135 ns |
|                List_Add |    61.24 ns |     7.991 ns |  0.438 ns |
|              List_Equal | 2,536.95 ns | 1,124.775 ns | 61.653 ns |
|           List_NotEqual | 2,416.80 ns |   273.405 ns | 14.986 ns |
| Dict_AddBinKeyFromEmpty |   665.40 ns |   189.475 ns | 10.386 ns |
| Dict_AddTxtKeyFromEmpty |   662.04 ns |    21.364 ns |  1.171 ns |
|             Dict_AddKey | 1,987.12 ns |    36.047 ns |  1.976 ns |
|        Dict_LookUpExist |    47.29 ns |    19.495 ns |  1.069 ns |
|     Dict_LookUpNonExist |    32.03 ns |     1.824 ns |  0.100 ns |
|              Dict_Equal | 7,375.63 ns | 1,504.051 ns | 82.442 ns |
|           Dict_NotEqual | 2,595.29 ns |    61.605 ns |  3.377 ns |

### Benchmarks after this patch

|           Method |             DataFile |      Mean |    Error |    StdDev |
|----------------- |--------------------- |----------:|---------:|----------:|
|  EncodeIntoBytes | /tmp/(...)e.dat [96] | 27.946 ms | 6.385 ms | 0.3500 ms |
| EncodeIntoStream | /tmp/(...)e.dat [96] | 21.738 ms | 3.136 ms | 0.1719 ms |
|  DecodeFromBytes | /tmp/(...)e.dat [96] |  8.232 ms | 4.241 ms | 0.2325 ms |
| DecodeFromStream | /tmp/(...)e.dat [96] |  8.601 ms | 7.952 ms | 0.4359 ms |

|                  Method |         Mean |      Error |    StdDev |
|------------------------ |-------------:|-----------:|----------:|
|    List_AddBinFromEmpty |    40.849 ns |  2.2660 ns | 0.1242 ns |
|    List_AddTxtFromEmpty |    40.057 ns |  0.6456 ns | 0.0354 ns |
|                List_Add |    50.728 ns |  8.4574 ns | 0.4636 ns |
|              List_Equal | 1,068.539 ns | 63.0099 ns | 3.4538 ns |
|           List_NotEqual | 1,027.934 ns | 60.0094 ns | 3.2893 ns |
| Dict_AddBinKeyFromEmpty |    73.809 ns |  1.4620 ns | 0.0801 ns |
| Dict_AddTxtKeyFromEmpty |    73.608 ns | 10.8463 ns | 0.5945 ns |
|             Dict_AddKey |   235.382 ns |  1.7675 ns | 0.0969 ns |
|        Dict_LookUpExist |    34.646 ns |  0.9301 ns | 0.0510 ns |
|     Dict_LookUpNonExist |    80.287 ns | 12.4760 ns | 0.6839 ns |
|              Dict_Equal | 2,369.966 ns | 92.6779 ns | 5.0800 ns |
|           Dict_NotEqual |     9.185 ns |  0.1734 ns | 0.0095 ns |

[1]: https://docs.microsoft.com/en-us/dotnet/api/system.collections.immutable.immutablesorteddictionary-2?view=netcore-3.1